P13

+++ backend/pandanews/src/main/java/sg/edu/smu/cs203/pandanews/controller/WorkGroupController.java
@@ -14,10 +14,17 @@ import org.springframework.web.bind.annotation.RestController;
 
 import sg.edu.smu.cs203.pandanews.service.WorkGroupService;
 import sg.edu.smu.cs203.pandanews.model.WorkGroup;
+import sg.edu.smu.cs203.pandanews.service.UserService;
+import sg.edu.smu.cs203.pandanews.repository.UserRepository;
+import sg.edu.smu.cs203.pandanews.model.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.context.SecurityContextHolder;
 
 @RestController
 public class WorkGroupController {
     private WorkGroupService workGroupService;
+    private UserService userService;
+    private UserRepository users;
 
     public WorkGroupController(WorkGroupService workGroupService){
         this.workGroupService = workGroupService;
@@ -27,9 +34,18 @@ public class WorkGroupController {
      * List all books in the system
      * @return list of all books
      */
-    @GetMapping("/workgroups")
-    public List<WorkGroup> getWorkGroups(){
-        return workGroupService.listWorkGroups();
+    @GetMapping("/organisations/{oid}/workgroups")
+    public List<WorkGroup> getWorkGroups(@PathVariable Long oid){
+        final UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        User user = userService.getUserByUsername(userDetails.getUsername());
+        if(user == null) return null;
+        // throw new 401
+
+        if(!users.findByOrganisationId(oid).contains(user)) return null;
+        // throw new 403
+
+        return workGroupService.listWorkGroups(oid);
     }
 
     /**
@@ -38,8 +54,17 @@ public class WorkGroupController {
      * @param id
      * @return book with the given id
      */
-    @GetMapping("/workgroups/{id}")
-    public WorkGroup getWorkGroup(@PathVariable Long id){
+    @GetMapping("/organisations/{oid}/workgroups/{id}")
+    public WorkGroup getWorkGroup(@PathVariable Long oid, @PathVariable Long id){
+        final UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        User user = userService.getUserByUsername(userDetails.getUsername());
+        if(user == null) return null;
+        // throw new 401
+
+        if(!users.findByOrganisationId(oid).contains(user)) return null;
+        // throw new 403
+
         WorkGroup workGroup = workGroupService.getWorkGroup(id);
 
         // Need to handle "book not found" error using proper HTTP status code
@@ -55,8 +80,17 @@ public class WorkGroupController {
      * @return list of all books
      */
     @ResponseStatus(HttpStatus.CREATED)
-    @PostMapping("/workgroups")
-    public WorkGroup addWorkGroup(@RequestBody WorkGroup workGroup){
+    @PostMapping("/organisations/{oid}/workgroups")
+    public WorkGroup addWorkGroup(@PathVariable Long oid, @RequestBody WorkGroup workGroup){
+        final UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        User user = userService.getUserByUsername(userDetails.getUsername());
+        if(user == null) return null;
+        // throw new 401
+
+        if(!users.findByOrganisationId(oid).contains(user)) return null;
+        // throw new 403
+
         return workGroupService.addWorkGroup(workGroup);
     }
 
@@ -66,8 +100,17 @@ public class WorkGroupController {
      * @param newOrganisationInfo
      * @return the updated, or newly added book
      */
-    @PutMapping("/workgroups/{id}")
-    public WorkGroup updateWorkGroup(@PathVariable Long id, @RequestBody WorkGroup newWorkGroupInfo){
+    @PutMapping("/organisations/{oid}/workgroups/{id}")
+    public WorkGroup updateWorkGroup(@PathVariable Long oid, @PathVariable Long id, @RequestBody WorkGroup newWorkGroupInfo){
+        final UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        User user = userService.getUserByUsername(userDetails.getUsername());
+        if(user == null) return null;
+        // throw new 401
+
+        if(!users.findByOrganisationId(oid).contains(user)) return null;
+        // throw new 403
+
         WorkGroup workGroup = workGroupService.updateWorkGroup(id, newWorkGroupInfo);
         if(workGroup == null) return null;
         
@@ -79,12 +122,21 @@ public class WorkGroupController {
      * If there is no book with the given "id", throw a BookNotFoundException
      * @param id
      */
-    @DeleteMapping("/workgroups/{id}")
-    public void deleteWorkGroup(@PathVariable Long id){
+    @DeleteMapping("/organisations/{oid}/workgroups/{id}")
+    public void deleteWorkGroup(@PathVariable Long oid, @PathVariable Long id){
+        final UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        User user = userService.getUserByUsername(userDetails.getUsername());
+        if(user == null) return;
+        // throw new 401
+
+        if(!users.findByOrganisationId(oid).contains(user)) return;
+        // throw new 403
+
         try {
             workGroupService.deleteWorkGroup(id);
         } catch(EmptyResultDataAccessException e) {
-            // throw new BookNotFoundException(id);
+            // throw new WorkGroupNotFoundException(id);
         }
     }
 }
